requirements.txt
python-dotenv
openai
yt-dlp
git+https://github.com/openai/whisper.git
